testing first circle as independent function

+++ src/interests/views.py
@@ -42,6 +42,9 @@ def all_interests(request):
 
 def save_interest(request, interest_id):
 	interest = Interest.objects.get(id=interest_id)
+	num_of_interests = Interest.objects.filter(userinterestanswer__user=request.user).count()
+	if num_of_interests >= 10:
+		return HttpResponseRedirect(reverse('interests'))
 	try: 
 		answered = UserInterestAnswer.objects.get(user=request.user, interest=interest)
 		answered.delete()
@@ -137,78 +140,6 @@ def all_interests(request):
 
 
 @user_passes_test(user_not_new)
-def edit_interests(request):
-
-	interests_all = Interest.objects.filter(userinterestanswer__user=request.user)
-	paginator = Paginator(interests_all, 1)
-	importance_levels = ['Strongly Dislike', 'Dislike', 'Neutral', 'Like', 'Strongly Like']
-	page = request.GET.get('page')
-
-	try:
-		interests = paginator.page(page)
-	except PageNotAnInteger:
-		#If page is not an integer, deliver first page.
-		interests = paginator.page(1)
-	except EmptyPage:
-		#If page is out of range, deliver last page of results
-		interests = paginator.page(paginator.num_pages)
-		#page = request.GET.get('page')
-		
-		#interest = interests.object_list[0]
-		#print "The Interest is: ", interest
-		#useranswer = UserInterestAnswer.objects.get(user=request.user, interest=interest)
-		#importance_level = useranswer.importance_level
-		#print "The importance level: ", useranswer.importance_level
-	if request.method == 'POST':
-		interest_id = request.POST['interest_id']
-
-		#user answer
-		importance_level = request.POST['importance_level']
-
-		user = User.objects.get(id=request.user.id)
-		interest = Interest.objects.get(id=interest_id)
-		'''
-		try:
-			interest_pic = InterestPicture.objects.get(interest=interest).filter(id=1)
-		except: 
-			pass
-		'''
-
-		#user answer save
-
-		answered, created = UserInterestAnswer.objects.get_or_create(user=user, interest=interest)
-		answered.importance_level = importance_level
-		answered.save()
-		return HttpResponseRedirect('')
-
-		
-	return render_to_response('interests/edit.html', locals(), context_instance=RequestContext(request))
-
-
-# displays the interests for a particular user
-@user_passes_test(user_not_new)
-def single_user_interests(request, username):
-	single_user = User.objects.get(username=username)
-	interests_all = Interest.objects.filter(userinterestanswer__user=single_user)
-	paginator = Paginator(interests_all, 1)
-	importance_levels = ['Strongly Dislike', 'Dislike', 'Neutral', 'Like', 'Strongly Like']
-
-	try: 
-		page = request.GET.get('page')
-		interests = paginator.page(page)
-	except PageNotAnInteger:
-		#If page is not an integer, deliver first page.
-		interests = paginator.page(1)
-	except EmptyPage:
-		#If page is out of range, deliver last page of results
-		interests = paginator.page(paginator.num_pages)
-	
-
-	return render_to_response('interests/single_user.html', locals(), context_instance=RequestContext(request))
-
-
-
-@user_passes_test(user_not_new)
 def search_interests(request):
 	try:
 		q = request.GET.get('q', '')
@@ -218,16 +149,4 @@ def search_interests(request):
 	results = interest_queryset
 	return render_to_response('interests/search.html', locals(), context_instance=RequestContext(request))
 
-
-@user_passes_test(user_not_new)
-def search_user_interests(request, username):
-	try:
-		q = request.GET.get('q', '')
-	except: 
-		q = False
-	interest_queryset = Interest.objects.filter(
-		Q(interest__icontains=q)
-		).filter(userinterestanswer__user__username=single_user)
-	results = interest_queryset
-	return render_to_response('interests/search.html', locals(), context_instance=RequestContext(request))
